% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsInmf.R
\name{assignClusterId}
\alias{assignClusterId}
\title{Map sample clusters to the correct subtype.
Computes the average expression value for each sample cluster and each signature.
Then iteratively assigns subtype IDs to the clusters starting with the cluster
with the highest certainty for a given subtype.
The function assumes that all samples and features are contained in the expression
matrix. Missing values will be ignored.}
\usage{
assignClusterId(samples, signatures, exprs)
}
\arguments{
\item{samples}{named list with sample clusters}

\item{exprs}{expresssion matrix with samples in columns and features in rows}

\item{named}{list of signatures}
}
\value{
the mapping between cluster ID and subtype label
}
\description{
Map sample clusters to the correct subtype.
Computes the average expression value for each sample cluster and each signature.
Then iteratively assigns subtype IDs to the clusters starting with the cluster
with the highest certainty for a given subtype.
The function assumes that all samples and features are contained in the expression
matrix. Missing values will be ignored.
}
\author{
Andreas Schlicker
}
