% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringTools.R
\name{averageSilhouetteWidth}
\alias{averageSilhouetteWidth}
\title{Calculates the average silhouette width of a clustering.}
\usage{
averageSilhouetteWidth(sw)
}
\arguments{
\item{clustering}{- a clustering object, such as the output of iNMF()}
}
\value{
avesw - a number: the average silhouette width
}
\description{
Calculates the average silhouette width of a clustering.
}
\examples{
print('Examples not written yet')
}
