% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsInmf.R
\name{clust2list}
\alias{clust2list}
\title{Convert list of clusters to a character vector}
\usage{
clust2list(clust)
}
\arguments{
\item{clust}{- list of clusters}
}
\value{
a character vector representation of the clusters (names are cluster members)
}
\description{
Convert list of clusters to a character vector
}
\examples{
print('Examples not written yet')
}
