% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{CMScolors}
\alias{CMScolors}
\title{Hard-code the CMS subtype colors, so that they can be reused easily.}
\usage{
CMScolors()
}
\arguments{
\item{NULL}{- no parameters}
}
\value{
- vector of 4 colors, shades of orange, blue, red and green
}
\description{
Hard-code the CMS subtype colors, so that they can be reused easily.
}
\examples{
print(CMScolors)
}
