% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{combatToGSE13294_GSE14333}
\alias{combatToGSE13294_GSE14333}
\title{M-combat with the GSE13294 and the GSE14333 datasets as reference.}
\usage{
combatToGSE13294_GSE14333(mat, plot = FALSE)
}
\arguments{
\item{mat}{- the matrix to be normalized to GSE13294/GSE14333}
}
\value{
a list of two variables: the normalized dataset alone, and combined with GSE13294/GSE14333
}
\description{
M-combat with the GSE13294 and the GSE14333 datasets as reference.
}
\examples{
org = loadPilotOrganoidExpression(meanCenter=FALSE, Log=TRUE)
orgT = org[, grepl('t', colnames(org))]
out = combatToGSE13294_GSE14333(orgT)
print(names(out))
print(head(out$alone))
}
