% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{combatToRef}
\alias{combatToRef}
\title{Normalize a given dataset to a reference dataset, using M-ComBat.
Source: https://github.com/aeolianine/M-ComBat/blob/master/MComBatRScript.R
Reference: Stein et al, Removing batch effects from purified plasma cell gene expression microarrays with modified ComBat,
           BMC Bioinformatics2015 16:63, https://doi.org/10.1186/s12859-015-0478-3}
\usage{
combatToRef(myDataset, refDataset, plot = FALSE)
}
\arguments{
\item{myDataset}{- the given dataset to be normalized}

\item{refDataset}{- the reference dataset to "be normalized to"}
}
\value{
a matrix corresponding to the normalized "myDataset"
        together with "refDataset" (pasted together column-wise)
}
\description{
Normalize a given dataset to a reference dataset, using M-ComBat.
Source: https://github.com/aeolianine/M-ComBat/blob/master/MComBatRScript.R
Reference: Stein et al, Removing batch effects from purified plasma cell gene expression microarrays with modified ComBat,
           BMC Bioinformatics2015 16:63, https://doi.org/10.1186/s12859-015-0478-3
}
\examples{
print('combatToRef(): The examples are not written yet!')
}
