% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{combatToTCGA}
\alias{combatToTCGA}
\title{M-combat with the TCGA COAD RNA-seq dataset as reference.}
\usage{
combatToTCGA(mat, plot = FALSE)
}
\arguments{
\item{mat}{- the matrix to be normalized to TCGA, with entries in log space}
}
\value{
a list of two variables: the normalized dataset alone, and combined with TCGA
}
\description{
M-combat with the TCGA COAD RNA-seq dataset as reference.
}
\examples{
org = loadPilotOrganoidExpression(meanCenter=FALSE, Log=TRUE)
orgT = org[, grepl('t', colnames(org))]
orgT = combatToTCGA(orgT)
print(head(orgT))
}
