% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{combineClassifiers}
\alias{combineClassifiers}
\title{Combining the three main classifiers with some variations into a consensus table. [This has to be revised given the new normalization]}
\usage{
combineClassifiers(matR, plot = FALSE, normalize = TRUE)
}
\arguments{
\item{matR}{- the input gene expression matrix, genes by samples; the non-mean-centered version, but logged (for RF and SSP)}

\item{plotTGFB}{- boolean, whether to plot TGFB expression or not, annotated by subtype}
}
\value{
Table - table of samples by classifiers, populated by the corresponding subtypes, including a consensus column
}
\description{
Combining the three main classifiers with some variations into a consensus table. [This has to be revised given the new normalization]
}
\examples{
orgM = loadPilotOrganoidExpression(meanCenter=TRUE)
orgM = orgM[, grepl('t', colnames(orgM))]
orgR = loadPilotOrganoidExpression(meanCenter=FALSE)
orgR = orgR[, grepl('t', colnames(orgR))]
res = combineClassifiers(orgR)
}
