% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{compare_RF_to_SSP}
\alias{compare_RF_to_SSP}
\title{Print quick comparison stats on the consensus between the RF and SSP classifiers
This function uses the results of subtypeCMS.RF and subtypeCMS.SSP}
\usage{
compare_RF_to_SSP(resRF, resSSP)
}
\arguments{
\item{resRF}{- the results data frame output of subtypeCMS.RF}

\item{resSSP}{- the results data frame output of subtypeCMS.SSP}
}
\value{
table (data.frame) of samples that disagree in assignment
        with their subtype and probabilities of assignment
}
\description{
Print quick comparison stats on the consensus between the RF and SSP classifiers
This function uses the results of subtypeCMS.RF and subtypeCMS.SSP
}
\examples{
mat = get_TCGA_synapse()
resRF = subtypeCMS.RF(mat)
resSSP = subtypeCMS.SSP(mat)
compare_RF_to_SSP(resRF, resSSP)
}
