% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{compare_to_training_labels}
\alias{compare_to_training_labels}
\title{Compare the subtypeCMS.RF assigned nearest subtypes to the RF model (gold reference) training labels.
This is only possible for datasets that are used within the reference set.
Note: Currently only TCGA is implemented.}
\usage{
compare_to_training_labels(resRF, whichDataset = NULL)
}
\arguments{
\item{resRF}{- output of subtypeCMS.RF}

\item{whichDataset}{- currently only "TCGA" is a valid input}
}
\value{
table (data.frame) of subtypes, training labels and posterior probabilities for samples that do not have matching labels;
        also a message with the percentage of matching sample labels
}
\description{
Compare the subtypeCMS.RF assigned nearest subtypes to the RF model (gold reference) training labels.
This is only possible for datasets that are used within the reference set.
Note: Currently only TCGA is implemented.
}
\examples{
mat = get_TCGA_synapse()
resRF = subtypeCMS.RF(mat)
compare_to_training_labels(resRF, whichDataset = 'TCGA')
}
