% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsInmf.R
\name{compareTwoClusteringsINMF}
\alias{compareTwoClusteringsINMF}
\title{Computes the number of mis-classified samples between two clusterings.}
\usage{
compareTwoClusteringsINMF(resInmf1, resInmf2, samplesOfInterest,
  whichStep = "step2")
}
\arguments{
\item{resInmf1}{- clustering object for the first clustering}

\item{resInmf2}{- clustering object for the second clustering}

\item{samplesOfInterest}{- out of the samples in the data, the subset of samples for which the comparison is performed}

\item{whichStep}{- which step in the clustering is being compared, 'step1' or 'step2', default is 'step2'}
}
\value{
list -
}
\description{
This is implemented for results of the iNMF algorithm, and could be ran for step1 or step2.
}
\examples{
print('Examples not written yet')
}
