% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsSadanandam.R
\name{compareTwoClusteringsSadanandam}
\alias{compareTwoClusteringsSadanandam}
\title{Computes the number of mis-classified samples between two clusterings.}
\usage{
compareTwoClusteringsSadanandam(res1, res2, samplesOfInterest)
}
\arguments{
\item{samplesOfInterest}{- out of the samples in the data, the subset of samples for which the comparison is performed}

\item{resInmf1}{- first clustering object to be compared}

\item{resInmf2}{- second clustering object to be compared}
}
\value{
list -
}
\description{
This is implemented for results of the Sadanandam algorithm.
}
