% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{convertRownamesToEZID}
\alias{convertRownamesToEZID}
\title{Convert gene symbol to Entrez ID}
\usage{
convertRownamesToEZID(geneNames)
}
\arguments{
\item{character}{vector of gene symbols}
}
\value{
character vector of entrez ids
}
\description{
Convert gene symbol to Entrez ID
}
\examples{
....
}
