% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsSadanandam.R
\name{createClusteringDataStructureSadanandam}
\alias{createClusteringDataStructureSadanandam}
\title{Convert a simple classes character vector to a clustering object.}
\usage{
createClusteringDataStructureSadanandam(whichSamples, clust,
  distance = "spearman")
}
\arguments{
\item{whichSamples}{- data (columns/samples) to which the clustering applies}

\item{clust}{- character vector of classes, names(clust) are the samples}

\item{distance}{- what is the distance metric for the clustering}

\item{signature}{- default is loadAndreasSignature()}
}
\value{
clustering object with silhouette widths
}
\description{
This produces an output similar to the iNMF() function output. This applies to the Sadanandam clusters only.
}
\examples{
...
}
