% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringTools.R
\name{distanceFunction}
\alias{distanceFunction}
\title{Distance function for two vectors. Auxiliary function.}
\usage{
distanceFunction(x, y, metric = "spearman")
}
\arguments{
\item{x}{- first vector}

\item{y}{- second vector}

\item{metric}{- distance metric, could be 'cor' for 1-pearson correlation or 'eucl' for Euclidean distance}
}
\value{
the distance between the two vector [numeric, length=1]
}
\description{
Distance function for two vectors. Auxiliary function.
}
\examples{
print('Examples not written yet')
}
