% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{entrez2symbol}
\alias{entrez2symbol}
\title{Convert Entrez ID to gene symbols}
\usage{
entrez2symbol(geneids)
}
\arguments{
\item{character}{vector of entrez ids}
}
\value{
character vector of gene symbols
}
\description{
Convert Entrez ID to gene symbols
}
\examples{
....
}
