% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{geneSymbol2EZID}
\alias{geneSymbol2EZID}
\title{This function is used within convertRownamesToEZID.}
\usage{
geneSymbol2EZID(geneName, dict = map)
}
\arguments{
\item{geneName}{- this is a character, the gene name to be translated.}

\item{dict}{- this is the map; the values are the desired translation, the names(dict) are the geneName(s)}
}
\value{
- the value of dict[ind], such that names(dict)[ind] = geneName
}
\description{
This function is used within convertRownamesToEZID.
}
\examples{
dict = c('a','b','c')
names(dict) = c('1','2','3')
geneSymbol2EZID('2',dict=dict)   # should return 'b'
}
