% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_GSE33113_synapse}
\alias{get_GSE33113_synapse}
\title{Downloads the frozen-RMA-normalized of the GSE33113 expression from Synapse.
Note 1: synapse ID "syn2363559"
Note 2: requires synapse login, can be automated by creating a .synapseConfig file.}
\usage{
get_GSE33113_synapse()
}
\value{
data frame of the expression matrix for the GSE33113 dataset
}
\description{
Downloads the frozen-RMA-normalized of the GSE33113 expression from Synapse.
Note 1: synapse ID "syn2363559"
Note 2: requires synapse login, can be automated by creating a .synapseConfig file.
}
\examples{
exprs = get_GSE33113_synapse()
print(dim(exprs))
print(exprs[1:3,1:3])
}
\seealso{
\code{\link{getSynapseTable}}
}
