% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_TCGA_synapse}
\alias{get_TCGA_synapse}
\title{Download the merged (Illumina -HiSeq and -GA ) TCGA expression data from Synapse.
Note 1: synapse ID "syn2325328"
Note 2: requires synapse login, can be automated by creating a .synapseConfig file.}
\usage{
get_TCGA_synapse()
}
\value{
data frame of the expression data, genes are rownames, samples as colnames
}
\description{
Download the merged (Illumina -HiSeq and -GA ) TCGA expression data from Synapse.
Note 1: synapse ID "syn2325328"
Note 2: requires synapse login, can be automated by creating a .synapseConfig file.
}
\examples{
tcga_expr = get_TCGA_synapse()
print(dim(tcga_exprs))
print(tcga_exprs[1:3,1:4])
}
\seealso{
\code{\link{getSynapseTable}}
}
