% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getSynapseTable}
\alias{getSynapseTable}
\title{This functions shows how to download data from Synapse.
The example is tailored to table-format data: an expression matrix
Note: requires synapse login, can be automated by creating a .synapseConfig file
      with the following contents:
      [authentication]
      username: username
      password: password}
\usage{
getSynapseTable(synapseID, sep = "\\t")
}
\arguments{
\item{synapseID}{- the synapse identifier associated with that dataset}

\item{sep}{- the table separator, ex: '\t', ','; default is tab}
}
\value{
data frame of the requested dataset
}
\description{
This functions shows how to download data from Synapse.
The example is tailored to table-format data: an expression matrix
Note: requires synapse login, can be automated by creating a .synapseConfig file
      with the following contents:
      [authentication]
      username: username
      password: password
}
\examples{
tab = getSynapseTable('syn2363559')
print(tab[1:5,1:3])
}
\seealso{
\code{\link{get_GSE33113_synapse}}
}
