% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsInmf.R
\name{iNMF}
\alias{iNMF}
\title{Run iNMF subtyping of the specified data set. This functions rests primarily on the subtype() function.}
\usage{
iNMF(exprs, signatures, types = NULL, bootstrap = FALSE,
  randomizeFeatures = FALSE, silhouette = TRUE, plotHeatmaps = TRUE,
  directory = ".", filePrefix = "", distance = "spearman")
}
\arguments{
\item{exprs}{expression matrix with samples in columns and features in rows}

\item{signatures}{named list with subtyping signatures mapped to feature ids
used in the expression matrix. The following names have to be used "step1",
"step2.c1" and "step2.c2". The function "signaturesList" can be used to
obtain this list.}

\item{bootstrap}{boolean indicating whether running in bootstrap mode. If TRUE, samples will be randomly selected}

\item{randomizeFeatures}{boolean indicating if features should be randomized before running the subtyping; default: FALSE}

\item{silhouette}{boolean indicating whether the silhouette should be calculated}

\item{plotHeatmaps}{boolean indicating whether clustering heatmaps should be saved in files}

\item{directory}{path to the directory where heatmaps will be saved}

\item{filePrefix}{file name prefix for the heatmap files}

\item{distance}{- what is the distance metric for the clustering}

\item{types:}{an annotation vector of size 1xncol(exprs)}
}
\value{
a named list with the clustering results
}
\description{
Run iNMF subtyping of the specified data set. This functions rests primarily on the subtype() function.
}
\author{
Andreas Schlicker, modified by Gergana Bounova
}
