% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inmf_signatures_entrez}
\alias{inmf_signatures_entrez}
\title{iNMF (Schlicker) subtypes signatures.}
\format{
List of 7 character vectors. Each vector contains entrez gene identifiers.
\describe{
  \item{1}{entrez identifiers of signature genes for subtype 1}
  \item{2}{entrez identifiers of signature genes for subtype 2}
  \item{1.1}{entrez identifiers of signature genes for subtype 1.1}
  \item{1.2}{entrez identifiers of signature genes for subtype 1.2}
  \item{1.3}{entrez identifiers of signature genes for subtype 1.3}
  \item{2.1}{entrez identifiers of signature genes for subtype 2.1}
  \item{2.2}{entrez identifiers of signature genes for subtype 2.2}
  ...
}
}
\source{
\url{https://bmcmedgenomics.biomedcentral.com/articles/10.1186/1755-8794-5-66}
}
\usage{
inmf_signatures_entrez
}
\description{
The genes defining each of the 5 subtypes in the Schlicker (iNMF) signature.
}
\keyword{datasets}
