% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inmf_signatures_sym_from_entrez}
\alias{inmf_signatures_sym_from_entrez}
\title{iNMF (Schlicker) subtypes signatures.}
\format{List of 7 character vectors. Each vector contains gene symbol identifiers
\describe{
  \item{1}{symbol identifiers of signature genes for subtype 1}
  \item{2}{symbol identifiers of signature genes for subtype 2}
  \item{1.1}{symbol identifiers of signature genes for subtype 1.1}
  \item{1.2}{symbol identifiers of signature genes for subtype 1.2}
  \item{1.3}{symbol identifiers of signature genes for subtype 1.3}
  \item{2.1}{symbol identifiers of signature genes for subtype 2.1}
  \item{2.2}{symbol identifiers of signature genes for subtype 2.2}
  ...
}}
\source{
\url{internal; to be edited}
}
\usage{
inmf_signatures_sym_from_entrez
}
\description{
The genes defining each of the 5 subtypes in the Schlicker (iNMF) signature.
These gene symbols were translated from the entrez ids (inmf_signatures_entrez),
     using the current version of biomaRt (June 11 2018).
}
\keyword{datasets}
