% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSignatures.R
\name{intersectSignature}
\alias{intersectSignature}
\title{Intersect a signature with the data's available genes}
\usage{
intersectSignature(genes, sig, print = TRUE)
}
\arguments{
\item{genes}{- the genes available in the data, usually rownames(data)}

\item{print}{- logical: print how many genes are in the intersection or not, default is TRUE}

\item{sigs}{- a list of the genes in every subtype (according to this signature)}
}
\value{
intsig - the "intersection" signature, comprising of only genes available in "genes"
}
\description{
Produces a signature with a same subtypes, but smaller sets of genes, corresponding to what is available in the data.
}
\examples{
intsig = intersectSignature(rownames(data),sig,print=TRUE)
stopifnot(names(intsig)==names(sig))
print(length(intsig))
}
