% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_signatures.R
\name{load_CMS_genes}
\alias{load_CMS_genes}
\title{Load random-forest genes from CMS classifier,
stored in CMSclassifier::finalModel$importance, with their entrez identifiers}
\usage{
load_CMS_genes(geneSymbol = FALSE)
}
\arguments{
\item{geneSymbol, }{boolean: should entrez identifiers be converted to gene symbols; default is FALSE}
}
\value{
a matrix with CMS genes as rownames, and random forest importance scores per subtype;
        this is basically "CMSclassifier::finalModel$importance", with row names converted to
        gene symbols if geneSymbol=TRUE
}
\description{
Load random-forest genes from CMS classifier,
stored in CMSclassifier::finalModel$importance, with their entrez identifiers
}
\examples{
mat = load_CMS_genes(geneSymbol=FALSE)
print(head(mat))
}
\seealso{
\code{\link{load_schlicker_signature}} and \code{\link{load_sadanandam_signature}}
}
