% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_signatures.R
\name{load_sadanandam_signature}
\alias{load_sadanandam_signature}
\title{Loading Sadanandam's subtype signatures.}
\usage{
load_sadanandam_signature()
}
\arguments{
\item{NULL, }{no inputs}
}
\value{
data frame of 786 genes by 5 subtypes; entries are relative expression values in the 5 subtype centroids
}
\description{
Reference: Sadanandam et al (2013). A colorectal cancer classification system that associates cellular phenotype and responses to therapy. Nature Medicine, 19(5), 619–25. doi:10.1038/nm.3175.
Note: In the paper these genes are refered to as "CRCassigner-786"
}
\examples{
sigs = load_sadanandam_signature()
print(sigs$Inflammatory)
}
\seealso{
\code{\link{load_schlicker_signature}} and \code{\link{load_CMS_genes}}
}
