% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_signatures.R
\name{load_schlicker_signature}
\alias{load_schlicker_signature}
\title{Loading the Schlicker (iNMF) subtype signatures from the paper supplement.}
\source{
\url{goo.gl/fghKBY},\url{goo.gl/vCzAwf}, \url{goo.gl/C5FpoE}, \url{goo.gl/uwABi1},
                            \url{goo.gl/pvzt33}, \url{goo.gl/b1eGqo}, \url{goo.gl/S6UhM3}
}
\usage{
load_schlicker_signature(annotation = "symbol")
}
\arguments{
\item{annotation}{(character), can be gene 'symbol' or 'entrez' identifier; default is 'symbol'}
}
\value{
a list of signatures
}
\description{
Reference: Schlicker et al (2012). Subtypes of primary colorectal tumors correlate with response to targeted
           treatment in colorectal cell lines. BMC Medical Genomics, 5(1), 66. doi:10.1186/1755-8794-5-66.
}
\examples{
sig = load_schlicker_signature()
print(sig$'1.1')
}
\seealso{
\code{\link{load_sadanandam_signature}} and \code{\link{load_CMS_genes}}
}
