% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSignatures.R
\name{loadSadanandamSignature}
\alias{loadSadanandamSignature}
\title{Loading Sadanandam's subtype signatures.}
\usage{
loadSadanandamSignature()
}
\arguments{
\item{NULL, }{no inputs}
}
\value{
data frame of 786 genes by 5 subtypes; entries are relative expression values in the 5 subtype centroids
}
\description{
Reference: Sadanandam et al (2013). A colorectal cancer classification system that associates cellular phenotype and responses to therapy. Nature Medicine, 19(5), 619–25. doi:10.1038/nm.3175.
Note: In the paper these genes are refered to as "CRCassigner-786"
}
\examples{
sigs = loadSadanandamSignature()
print(sigs$Inflammatory)
}
\seealso{
\code{\link{loadSchlickerSignature}} which is another function loading signatures
}
