% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSignatures.R
\name{loadSadanandamSignatureDeprecated}
\alias{loadSadanandamSignatureDeprecated}
\title{Loading Sadanandam's subtype signatures as a list.}
\usage{
loadSadanandamSignatureDeprecated()
}
\arguments{
\item{no}{inputs needed}
}
\value{
a list of signatures
}
\description{
Reference: Sadanandam et al (2013). A colorectal cancer classification system that associates cellular phenotype and responses to therapy. Nature Medicine, 19(5), 619–25. doi:10.1038/nm.3175.
CRCassigner-786
}
\examples{
sigs = loadSadanandamSignature()
print(sigs$Inflammatory)
}
\seealso{
\code{\link{loadAndreasSignature}} which is another function loading signatures
}
