% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringTools.R
\name{nearestCentroid}
\alias{nearestCentroid}
\title{Implement a simple version of nearest-centroid classification.
Centroids can be computed from labels, or given directly.}
\usage{
nearestCentroid(df, samples = NULL, distance = "spearman",
  labels = NULL, centroids = NULL)
}
\arguments{
\item{df}{- a genes by samples matrix}

\item{samples}{- a single [new] sample to be labeled; if NULL, all columns of df are (re-)labeled; default is NULL}

\item{distance}{- vector distance, could be 'cor' (for 1-pearson correlation) or 'eucl' (for Euclidean)}

\item{labels}{- sample labels, could be NULL}

\item{centroids}{- pre-computed centroids, could be NULL}
}
\value{
corresponding labels for all samples
}
\description{
Implement a simple version of nearest-centroid classification.
Centroids can be computed from labels, or given directly.
}
\examples{
print('Examples not written yet')
}
