% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsSadanandam.R
\name{nearestCentroidsSadanandam}
\alias{nearestCentroidsSadanandam}
\title{Use the centroid matrix from the Sadanandam paper to assign samples to the nearest centroid.}
\usage{
nearestCentroidsSadanandam(df, types = NULL, plotHeatmaps = F,
  plotSilWidths = F, directory = ".", filePrefix = "",
  distance = "spearman")
}
\arguments{
\item{df}{- a data frame representing the data matrix}

\item{distance}{- what is the distance metric for the clustering}

\item{types:}{samples annotation (optional, character vector)}
}
\value{
final predicted classes for the samples in df
}
\description{
Using Euclidean distance for sample to centroid distance.
}
