% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{OLD_combatToGSE35896}
\alias{OLD_combatToGSE35896}
\title{M-combat with the GSE35896 dataset as reference.}
\usage{
OLD_combatToGSE35896(mat, plot = FALSE)
}
\arguments{
\item{mat}{- the matrix to be normalized to GSE35896}
}
\value{
a list of two variables: the normalized dataset alone, and combined with GSE35896
}
\description{
M-combat with the GSE35896 dataset as reference.
}
\examples{
org = loadPilotOrganoidExpression(meanCenter=FALSE, Log=TRUE)
orgT = org[, grepl('t', colnames(org))]
out = combatToGSE35896(orgT)
print(names(out))
print(head(out$alone))
}
