% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineClassifiers.R
\name{OLD_combineClassifiers}
\alias{OLD_combineClassifiers}
\title{Combining the three main classifiers with some variations into a consensus table. [This has to be revised given the new normalization]}
\usage{
OLD_combineClassifiers(matR, plot = FALSE, normalize = TRUE)
}
\arguments{
\item{matR}{- the input gene expression matrix, genes by samples; the non-mean-centered version, but logged (for RF and SSP)}

\item{plotTGFB}{- boolean, whether to plot TGFB expression or not, annotated by subtype}
}
\value{
Table - table of samples by classifiers, populated by the corresponding subtypes, including a consensus column
}
\description{
Combining the three main classifiers with some variations into a consensus table. [This has to be revised given the new normalization]
}
