% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{OLD_createHeatmap}
\alias{OLD_createHeatmap}
\title{Create subtypes heatmap.}
\usage{
OLD_createHeatmap(exprs, clustering, signatures, types = NULL,
  anno_colors = NULL, filename = NA, cellwidth = 1, cellheight = 1)
}
\arguments{
\item{exprs}{- expression matrix (genes by samples)}

\item{clustering}{- list of sample clusters}

\item{signatures}{- list of genes per subtype}

\item{types}{- factor for samples types, for example c('organoids', 'TCGA-RNAseq'), used as column annotation in aheatmap()}

\item{anno_colors}{- used for annotation colors in aheatmap(), default is NULL}

\item{filename}{- to be used within aheatmap(), default is NA}

\item{cellwidth/cellheight}{- inputs for aheatmap(), to control the size of cells plotted}
}
\description{
Adapted from Andreas's createHeatmap() function. Need to add examples and explain inputs. Uses the function aheatmap() from the NMF package.
}
\examples{
createHeatmap(mat,clust,intsig, types = src, filename = 'heatmap_organoids_rnaSeq_inmf_complete.png')
}
\seealso{
\code{\link[NMF]{aheatmap}}
}
