% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{OLD_plotClusterMDSSchlicker}
\alias{OLD_plotClusterMDSSchlicker}
\title{Plot two mds components for results summarized in a clustering object.}
\usage{
OLD_plotClusterMDSSchlicker(exprs, resInmf, types, method = "cmdscale",
  plotName = NULL, title = "")
}
\arguments{
\item{exprs}{- expression matrix}

\item{resInmf}{- a clustering object, the output of iNMF()}

\item{types}{- indicating different sample types (factor of level 2)}

\item{method}{- can be 'cmdscale', 'princomp', 'isoMDS', or 'tsne'; default is 'cmdscale'}

\item{plotName, }{if the plot is being saved; default is NULL}

\item{title}{- character string for the title of the plot, if any; default is ''}
}
\description{
Uses the iNMF type results and cmdscale, as well as "type" annotation; and the Schlicker signature.
}
