% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{OLD_plotSilhouetteWidths}
\alias{OLD_plotSilhouetteWidths}
\title{Plot silhouette widths.}
\usage{
OLD_plotSilhouetteWidths(sw)
}
\arguments{
\item{sw, }{silhouette width object}
}
\description{
Plot silhouette widths.
}
