% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{OLD_subtypeCMS.SSP}
\alias{OLD_subtypeCMS.SSP}
\title{Applying the single-sample classifier of the CMS consortium.}
\usage{
OLD_subtypeCMS.SSP(mat, plot = FALSE, plotWhichSamples = c())
}
\arguments{
\item{mat}{- the matrix to be subtyped, genes by samples, genes are typically in gene symbol format}

\item{plot}{- Boolean, TRUE or FALSE}
}
\value{
- table of subtype information, columns are predicted CMS, nearest CMS and poster
}
\description{
Applying the single-sample classifier of the CMS consortium.
}
