% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{plot_mCombat_effect}
\alias{plot_mCombat_effect}
\title{Visualize the effect that M-combat has on normalization.
Uses the input and results of the combatToRef()}
\usage{
plot_mCombat_effect(myDataset, refDataset, normDataset)
}
\arguments{
\item{myDataset}{- the original dataset to be normalized}

\item{refDataset}{- the reference dataset}

\item{normDataset}{- the normalized version of the original dataset}
}
\value{
plots - what kind of plots?
}
\description{
Visualize the effect that M-combat has on normalization.
Uses the input and results of the combatToRef()
}
\examples{
print('Examples not written yet')
}
