% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{plotClusterMDSSadanandam}
\alias{plotClusterMDSSadanandam}
\title{Plot two mds components for results summarized in a clustering object.}
\usage{
plotClusterMDSSadanandam(exprs, clust, types, method = "cmdscale",
  plotName = NULL, title = "")
}
\arguments{
\item{exprs}{- expression matrix}

\item{types}{- indicating different sample types (factor of level 2)}

\item{method}{- can be 'cmdscale', 'princomp', 'isoMDS', or 'tsne'; default is 'cmdscale'}

\item{plotName, }{if the plot is being saved; default is NULL}

\item{title}{- character string for the title of the plot, if any; default is ''}

\item{resInmf}{- a clustering object, the output of iNMF()}
}
\description{
Uses the Sadanandam type results and cmdscale, as well as "type" annotation; and the Sadanandam signature.
}
\examples{
...
}
