% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{plotSadanandamSubtypeCentroids}
\alias{plotSadanandamSubtypeCentroids}
\title{Plotting the scores from the signature table (CRCassigner-786) as a heatmap}
\usage{
plotSadanandamSubtypeCentroids()
}
\arguments{
\item{no}{inputs needed}
}
\description{
Reference: Sadanandam et al (2013). A colorectal cancer classification system that associates
     cellular phenotype and responses to therapy. Nature Medicine, 19(5), 619–25. doi:10.1038/nm.3175
}
\examples{
plotSadanandamSignatureScores()
}
\seealso{
\code{\link{loadSadanandamSignature}}
}
