% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{plotSilhouetteWidths}
\alias{plotSilhouetteWidths}
\title{Plot silhouette widths.}
\usage{
plotSilhouetteWidths(sw)
}
\arguments{
\item{sw, }{silhouette width object}
}
\description{
Plot silhouette widths.
}
\examples{
...
}
