% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{plotSubtypesCMS.RF}
\alias{plotSubtypesCMS.RF}
\title{Visualization of the results of subtypeCMS.RF().
   Creates barplot of posterior probabilities per sample.
   Also plots a heatmap of the expression of the random forest genes, using all samples.
   (This part has to be re-implemented. Pending on the multiFactorHeatmap rewrite.)}
\usage{
plotSubtypesCMS.RF(res)
}
\arguments{
\item{res}{- the result of subtypeCMS.RF()}
}
\description{
Visualization of the results of subtypeCMS.RF().
   Creates barplot of posterior probabilities per sample.
   Also plots a heatmap of the expression of the random forest genes, using all samples.
   (This part has to be re-implemented. Pending on the multiFactorHeatmap rewrite.)
}
