% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{plotSubtypesCMS.SSP}
\alias{plotSubtypesCMS.SSP}
\title{Visualization of the results of subtypeCMS.SSP().
   Creates barplot of assigned predicted and nearest subtypes.
   Also plots a heatmap of the computed correlations to all subtype centroids.}
\usage{
plotSubtypesCMS.SSP(res)
}
\arguments{
\item{res}{- the result of subtypeCMS.SSP()}
}
\value{
two plots, a barplot and a correlations heatmap
}
\description{
Visualization of the results of subtypeCMS.SSP().
   Creates barplot of assigned predicted and nearest subtypes.
   Also plots a heatmap of the computed correlations to all subtype centroids.
}
\examples{
mat = get_TCGA_synapse()
res = subtypeCMS.SSP(mat)
plotSubtypesCMS.SSP(res)
}
