% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringTools.R
\name{randClusteringMetric}
\alias{randClusteringMetric}
\title{Implement W. Rand's metric for comparing two clusters.}
\usage{
randClusteringMetric(numItems, clustering1, clustering2)
}
\arguments{
\item{numItems}{- total number of items being clustered}

\item{clustering1}{- list of clusters in clustering 1; each cluster is a vector of items}

\item{clustering2}{- list of clusters in clustering 2; each cluster is a vector of items}
}
\value{
Rand's metric as defined in his paper.
}
\description{
Source: Rand, W. M. (1971). Objective Criteria for the Evaluation of Clustering Methods. Journal of the American Statistical Association, 66(336), 846–850.
}
\examples{
c = randClusteringMetric(5, list(1=c(1,2,3), 2=c(4,5)), list(1=c(1), 2=c(2,3,4), 3=c(5)) )
print(c)
}
