% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sadanandam_786_genes}
\alias{sadanandam_786_genes}
\title{The 786 genes identified in the PAM analysis by Sadandandam et al, including the 5 subtype centroids.}
\format{
Data frame of 786 genes by 5 subtypes. Each entry represents the relative expression of that
        gene in the centroid of the corresponding subtype;
}
\source{
\url{https://images.nature.com/original/nature-assets/nm/journal/v19/n5/extref/nm.3175-S2.xls}
}
\usage{
sadanandam_786_genes
}
\description{
Source: Sadanandam et al (2013). A colorectal cancer classification system that associates cellular
         phenotype and responses to therapy. Nature Medicine, 19(5), 619–25. doi:10.1038/nm.3175.
         Suppl Table 1, "PAM" sheet
}
\keyword{datasets}
