% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsInmf.R
\name{shrunkenCentroidsINMF}
\alias{shrunkenCentroidsINMF}
\title{Apply the shrunken centroids method to assign a sample to a subtype. Use the iNMF signature here.}
\usage{
shrunkenCentroidsINMF(df, types = NULL, plotHeatmap = F, directory = ".",
  filePrefix = "", distance = "spearman")
}
\arguments{
\item{df}{- data frame/matrix of expression values (genes by samples)}

\item{distance}{- what is the distance metric for the clustering}

\item{types:}{samples annotation (optional, character vector)}
}
\value{
final predicted classes for the samples in df
}
\description{
Ref: Tibshirani, R. (2002). Diagnosis of multiple cancer types by shrunken centroids of gene expression. PNAS, 99(10), 6567–72.
}
\examples{
...
}
