% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSignatures.R
\name{signaturesGeneOverlap}
\alias{signaturesGeneOverlap}
\title{Comparing signatures: overlapping genes per subtype}
\usage{
signaturesGeneOverlap(sig1, sig2)
}
\arguments{
\item{sig1}{the first signature}

\item{sig2}{the second signature}
}
\value{
matrix of number of common genes per subtype
}
\description{
This function computes the overlapping genes per signature subtype for two signatures.
This is an indirect comparison of signature similarity since genes expression scores can be correlated.
}
\examples{
signaturesGeneOverlap(loadAndreasSignature(), loadSadanandamSignature())
}
