% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsInmf.R
\name{subtype}
\alias{subtype}
\title{Performs one step of the iNMF clustering. Essentially, this is a
hierarchical clustering of the given samples using the given signatures.
Cluster IDs are assigned using function "assignClusterId" but should be
manually verified using a heatmap or similar. The number of clusters is
equal to the number of signatures. Also computes silhouette values for the
clustering.}
\usage{
subtype(exprs, signatures, samples = NULL, silhouette = TRUE,
  distance = "spearman")
}
\arguments{
\item{exprs}{expression matrix with samples in columns and features in rows}

\item{signatures}{list of signatures to use.}

\item{samples}{vector with sample names. If NULL (default), all samples will be
clustered.}

\item{silhouette}{boolean value indicating whether the silhouette is to be
computeted}
}
\value{
Named list with two elements. The first element "clustering" contains a
named list with the assignment of samples to clusters. The names correspond to the
names used for the signatures. The second element "silhouette" contains an object
of class silhouette, if silhouette==TRUE.
}
\description{
Performs one step of the iNMF clustering. Essentially, this is a
hierarchical clustering of the given samples using the given signatures.
Cluster IDs are assigned using function "assignClusterId" but should be
manually verified using a heatmap or similar. The number of clusters is
equal to the number of signatures. Also computes silhouette values for the
clustering.
}
\author{
Andreas Schlicker
}
