% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{subtypeCMSOld}
\alias{subtypeCMSOld}
\title{Subtype samples using the temporary Sage-CMS classifier.}
\usage{
subtypeCMSOld(mat, impute = FALSE)
}
\arguments{
\item{mat}{- the input gene expression matrix, genes by samples (works better if gene-wise mean centered)}

\item{impute}{- Boolean, impute or not the missing genes}
}
\value{
res - factor of CMSx values {CMS1, CMS2, CMS3, CMS4}. The factor/vector names are the sample names (or colnames(mat))
}
\description{
Subtype samples using the temporary Sage-CMS classifier.
}
\examples{
....
}
