% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsSadanandam.R
\name{subtypeOneAtATimeSadanandam}
\alias{subtypeOneAtATimeSadanandam}
\title{Subtyping samples one at a time, using Sadanandam's subtypes.}
\usage{
subtypeOneAtATimeSadanandam(whichSamples, types = NULL,
  plotHeatmaps = FALSE, plotSilWidths = FALSE, method = "maxmean",
  directory = ".", filePrefix = "", distance = "spearman",
  silhouette = TRUE)
}
\arguments{
\item{whichSamples}{- a data frame (gene by samples) to be subtyped}

\item{plotHeatmaps}{- TRUE/FALSE, whether to create heatmap plots per subtype, default is FALSE}

\item{plotSilWidths}{- TRUE/FALSE, whether to create silhouette width plots per subtype, default is FALSE}

\item{method}{- default 'maxmean': assigning cluster expression means to signature subtype expression means}

\item{directory}{- string, naming the directory in which to save the heatmap/silhouette width plots, default is '.'}

\item{filePrefix}{- string, indicating a prefix to be added to the heatmap/sil width plot names, default is ''}

\item{distance}{- what is the distance metric for the clustering}
}
\value{
res - clustering and silhouette widths of the samples
}
\description{
Right now the implementation uses the subtype() function to assign 1 sample to a subtype. The subtypes are Inflammatory, TA, Stem.like, Goblet.like and Enterocyte.
}
\examples{
res = subtypeOneAtATimeSadanandam(organoidsT)
}
