% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsInmf.R
\name{subtypeSamplesINMF}
\alias{subtypeSamplesINMF}
\title{Run the iNMF (Schlicker) subtyping on a set of samples.}
\usage{
subtypeSamplesINMF(whichSamples, types = NULL, plotHeatmaps = FALSE,
  plotSilWidths = FALSE, directory = ".", filePrefix = "",
  distance = "spearman", silhouette = TRUE)
}
\arguments{
\item{whichSamples}{- a data frame (gene by samples) to be subtyped}

\item{plotHeatmaps}{- TRUE/FALSE, whether to create heatmap plots per subtype, default is FALSE}

\item{plotSilWidths}{- TRUE/FALSE, whether to create silhouette width plots per subtype, default is FALSE}

\item{directory}{- string, naming the directory in which to save the heatmap/silhouette width plots, default is '.'}

\item{filePrefix}{- string, indicating a prefix to be added to the heatmap/sil width plot names, default is ''}

\item{distance}{- what is the distance metric for the clustering}

\item{silhouette}{- Boolean; compute silhouette widths or not; default is TRUE}
}
\value{
resInmf - clustering and silhouette widths of the samples
}
\description{
Run the iNMF (Schlicker) subtyping on a set of samples.
}
\examples{
subtypeSamplesINMF(organoidsT)
subtypeSamplesINMF(cellLines, plotHeatmaps = T)
}
