% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsSadanandam.R
\name{subtypeSamplesSadanandam}
\alias{subtypeSamplesSadanandam}
\title{Subtype samples into Sadanandam's subtypes using Andreas's subtype() function.}
\usage{
subtypeSamplesSadanandam(whichSamples, types = NULL,
  plotHeatmaps = FALSE, plotSilWidths = FALSE, directory = ".",
  filePrefix = "", distance = "spearman", silhouette = TRUE)
}
\arguments{
\item{whichSamples}{- a data frame (gene by samples) to be subtyped}

\item{types}{- samples annotation (a character or numeric vector, size 1xncol(whichSamples))}

\item{plotHeatmaps}{- TRUE/FALSE, whether to create heatmap plots per subtype, default is FALSE}

\item{plotSilWidths}{- TRUE/FALSE, whether to create silhouette width plots per subtype, default is FALSE}

\item{directory}{- string, naming the directory in which to save the heatmap/silhouette width plots, default is '.'}

\item{filePrefix}{- string, indicating a prefix to be added to the heatmap/sil width plot names, default is ''}

\item{distance}{- what is the distance metric for the clustering}
}
\value{
res - clustering and silhouette widths of the samples
}
\description{
Subtype samples into Sadanandam's subtypes using Andreas's subtype() function.
}
