% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSignatures.R
\name{symbol2entrez}
\alias{symbol2entrez}
\title{Converts gene symbols to entrez ids using org.Hs.egSYMBOL2EG
         from the org.Hs.eg.db package.}
\usage{
symbol2entrez(geneNames)
}
\arguments{
\item{geneNames}{- character vector}
}
\value{
character vector, the entrez identifiers corresponding to geneName, or the geneName itself if it is not mapped
}
\description{
Converts gene symbols to entrez ids using org.Hs.egSYMBOL2EG
         from the org.Hs.eg.db package.
}
\examples{
print(symbol2entrez('KRAS'))
print(symbol2entrez('shouldnotbemapped'))
print(symbol2entrez(c('KRAS', 'AAAAAAAA', 'TP53', 'KRAS')))
}
