% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsInmf.R
\name{unpackINMF}
\alias{unpackINMF}
\title{Unpack the INMF step1, step2.c1, step2.c2 structure into a dataframe with assignment, neighbor and sil width}
\usage{
unpackINMF(resInmf)
}
\arguments{
\item{resInmf}{- [list] =, result of iNMF()}
}
\value{
data frame with cluster assignment, neighboring cluster and sil width
}
\description{
Unpack the INMF step1, step2.c1, step2.c2 structure into a dataframe with assignment, neighbor and sil width
}
\examples{
print('Examples not written yet')
}
