% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsInmf.R
\name{unpackINMFclustering}
\alias{unpackINMFclustering}
\title{Convert the INMF step1, step2.c1, step2.c2 structure to lists of clusters.}
\usage{
unpackINMFclustering(resInmf)
}
\arguments{
\item{resInmf}{- [list] =, result of iNMF()}
}
\value{
simple list of clusters
}
\description{
Convert the INMF step1, step2.c1, step2.c2 structure to lists of clusters.
}
\examples{
print('Examples not written yet')
}
