% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSignatures.R
\name{entrez2symbol}
\alias{entrez2symbol}
\title{Convert Entrez ID to gene symbols.
Note: uses the "org.Hs.eg.db" package.}
\usage{
entrez2symbol(geneids)
}
\arguments{
\item{character}{vector of entrez ids}
}
\value{
character vector of gene symbols, with names as entrez ids
}
\description{
Convert Entrez ID to gene symbols.
Note: uses the "org.Hs.eg.db" package.
}
\examples{
print(c('KRAS', 'NRAS', 'BRAF', 'APC', 'COL3A1', NA) == entrez2symbol(c('3845', '4893', '673', '324', '1281', '00000')))
}
