% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSignatures.R
\name{intersectSignature}
\alias{intersectSignature}
\title{Intersect a signature with the data's available genes}
\usage{
intersectSignature(genes, sig, verbose = TRUE)
}
\arguments{
\item{genes}{- the genes available in the data, usually rownames(data)}

\item{verbose}{- logical: print how many genes are in the intersection or not, default is TRUE}

\item{sigs}{- a list of the genes in every subtype (according to this signature);
or a character vector of genes, if that signature is not split into subtypes}
}
\value{
intsig - the "intersection" signature, comprising of only signature genes available in "genes"
}
\description{
Produces a signature with the same subtypes, but smaller sets of genes, corresponding to what is available in the data.
}
\examples{
intersectSignature(c('ARID1B', 'PTEN'),c('KRAS', 'TGFB1', 'ARID1B'), verbose=TRUE)
}
