% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{subtypeCMS.RF}
\alias{subtypeCMS.RF}
\title{Applying the random forest classifier of the CMS consortium.}
\usage{
subtypeCMS.RF(mat, plot = FALSE)
}
\arguments{
\item{mat}{- the matrix to be subtyped, genes by samples, genes are typically in gene symbol format}

\item{plot}{- Boolean, TRUE or FALSE}
}
\value{
- table of subtype information, columns are predicted CMS, nearest CMS and posterior probabilities of CMS assignment
}
\description{
Applying the random forest classifier of the CMS consortium.
}
