% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{subtypeCMS.RF}
\alias{subtypeCMS.RF}
\title{Applying the random forest classifier of the CMS consortium (CMSclassifier::classifyCMS.RF)
Installing the CMS classifier: "library(devtools); install_github("Sage-Bionetworks/CMSclassifier")"}
\usage{
subtypeCMS.RF(mat)
}
\arguments{
\item{mat}{- the matrix to be subtyped, genes by samples, genes are typically represented as gene symbols}
}
\value{
- table of subtype information, columns are predicted CMS,
         nearest CMS and posterior probabilities of CMS assignment;
         rownames are samples
}
\description{
Applying the random forest classifier of the CMS consortium (CMSclassifier::classifyCMS.RF)
Installing the CMS classifier: "library(devtools); install_github("Sage-Bionetworks/CMSclassifier")"
}
\examples{
mat = get_TCGA_synapse()
res = subtypeCMS.RF(mat)
}
