% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypingToolsCMS.R
\name{subtypeCMS.SSP}
\alias{subtypeCMS.SSP}
\title{Applying the single-sample classifier of the CMS consortium (CMSclassifier::classifyCMS.SSP)
Installing the CMS classifier: "library(devtools); install_github("Sage-Bionetworks/CMSclassifier")"}
\usage{
subtypeCMS.SSP(mat)
}
\arguments{
\item{mat}{- the matrix to be subtyped, genes by samples, genes are typically in gene symbol format}
}
\value{
- table of samples and their predicted, nearest CMS assignment, as well as min/median and max
          correlations to predefined CMS1, CMS2, CMS3 and CMS4 centroids
}
\description{
Applying the single-sample classifier of the CMS consortium (CMSclassifier::classifyCMS.SSP)
Installing the CMS classifier: "library(devtools); install_github("Sage-Bionetworks/CMSclassifier")"
}
\examples{
mat = get_TCGA_synapse()
res = subtypeCMS.SSP(mat)
}
