% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_signatures.R
\name{symbol2entrez}
\alias{symbol2entrez}
\title{Convert Entrez ID to gene symbols.
Note 1: Uses the "org.Hs.eg.db" package.
Note 2: Deals with missing keys by returning NAs.}
\usage{
symbol2entrez(geneids)
}
\arguments{
\item{character}{vector of gene symbols}
}
\value{
character vector of entrez ids
}
\description{
Convert Entrez ID to gene symbols.
Note 1: Uses the "org.Hs.eg.db" package.
Note 2: Deals with missing keys by returning NAs.
}
\examples{
print(symbol2entrez(c('KRAS', 'NRAS', 'BRAF', 'APC', 'COL3A1', 'AAAAAAA')) == c('3845', '4893', '673', '324', '1281', NA))
}
